Moved exceptions classes to separate folder. Added speech service existence check, before call its functions. Added spech service destruction before program closing.
