+++ android/src/main/java/org/vosk/vosk_flutter_plugin/exceptions/SpeechServiceNotFound.java
@@ -1,0 +1,7 @@
+package org.vosk.vosk_flutter_plugin.exceptions;
+
+public class SpeechServiceNotFound extends Exception{
+  public SpeechServiceNotFound(){
+    super("Speech service not initialized");
+  }
+}
