+++ android/src/main/java/org/vosk/vosk_flutter_plugin/exceptions/WrongArgumentTypeException.java
@@ -1,4 +1,4 @@
-package org.vosk.vosk_flutter_plugin;
+package org.vosk.vosk_flutter_plugin.exceptions;
 
 public class WrongArgumentTypeException extends Exception {
